extract web, view
