+++ view/types.go
@@ -1,4 +1,4 @@
-package api
+package view
 
 import (
 	"changemedaddy/invest"
